refactor: extract graph methods from diagram.ts
