+++ docs/vocabulary.md
@@ -2,6 +2,8 @@
 
 This is a list of terms used for the project, whose meanings may not be obvious. Hopefully this list can remain short, but sometimes coming up with an intuitive name is hard.
 
+- graph: a set of nodes and edges
+- diagram: a graph or subset of a graph that is displayed (and therefore layout becomes relevant)
 - graph part: a node or edge
 - topic data: all graph parts and scores associated with a topic
 - topic diagram: the diagram that shows when you initially visit a topic
