+++ src/web/topic/components/CriteriaTable/CriteriaTable.tsx
@@ -18,8 +18,8 @@ import { useSessionUser } from "../../../common/hooks";
 import { useCriterionSolutionEdges, useNode, useNodeChildren } from "../../store/nodeHooks";
 import { useUserCanEditTopicData } from "../../store/userHooks";
 import { closeTable } from "../../store/viewActions";
-import { Edge, Node } from "../../utils/diagram";
 import { getConnectingEdge } from "../../utils/edge";
+import { Edge, Node } from "../../utils/graph";
 import { EdgeCell } from "../CriteriaTable/EdgeCell";
 import { NodeCell } from "../CriteriaTable/NodeCell";
 import { AddNodeButton } from "../Node/AddNodeButton";
