+++ src/web/topic/components/Diagram/Diagram.tsx
@@ -19,7 +19,8 @@ import { useViewportUpdater } from "../../hooks/flowHooks";
 import { setSelected } from "../../store/actions";
 import { connectNodes, reconnectEdge } from "../../store/createDeleteActions";
 import { useUserCanEditTopicData } from "../../store/userHooks";
-import { Diagram as DiagramData, type Edge, type Node } from "../../utils/diagram";
+import { Diagram as DiagramData } from "../../utils/diagram";
+import { type Edge, type Node } from "../../utils/graph";
 import { Orientation } from "../../utils/layout";
 import { FlowNodeType } from "../../utils/node";
 import { FlowEdge } from "../Edge/FlowEdge";
