+++ src/web/topic/components/Indicator/NodeIndicatorGroup.tsx
@@ -1,6 +1,6 @@
 import { Box } from "@mui/material";
 
-import { Node } from "../../utils/diagram";
+import { Node } from "../../utils/graph";
 import { CriteriaTableIndicator } from "../Indicator/CriteriaTableIndicator";
 import { NodeClaimIndicator } from "../Indicator/NodeClaimIndicator";
 import { Score } from "../Score/Score";
