+++ src/web/topic/utils/apiConversion.ts
@@ -8,7 +8,7 @@ import { Node as ApiNode, Node } from "../../../common/node";
 import { UserScore as ApiScore } from "../../../common/userScore";
 import { UserScores as StoreScores, TopicStoreState } from "../store/store";
 import { isPlaygroundTopic } from "../store/utils";
-import { Score, Edge as StoreEdge, Node as StoreNode, buildEdge, buildNode } from "./diagram";
+import { Score, Edge as StoreEdge, Node as StoreNode, buildEdge, buildNode } from "./graph";
 
 export const convertToStoreNode = (apiNode: TopicNode) => {
   return buildNode({
